chore: reverted naming, using hack to allow test suite to skip on non tfacc
